deprecate old training scripts and utils
