+++ conf/basic.yaml
@@ -1,17 +1,20 @@
-data_path: '/home/zhz/dataset/places365_standard'
-log_dir: logs
+data_path: /root/TorchScene/places365_standard/
+device: cuda:0
 weight_dir: checkpoints
-arch: mobilenet_v2
+arch: vision_transformer
 workers: 6
-epochs: 90
+epochs: 7
 start_epoch: 0
-batch_size: 256
-lr: 0.1
+batch_size: 128
+lr: 1e-4
 momentum: 0.9
-weight_decay: 1e-4
-print_freq: 10
-resume: False
-evaluate: False
-pretrained: False
+weight_decay: 1e-5
+print_freq: 100
+pretrained: True
+feature_extract: False
 num_classes: 365
 dataset: places365
+
+hydra:
+  run:
+    dir: /root/TorchScene/logs/finetune/${now:%Y-%m-%d}/${now:%H-%M-%S}
