+++ deprecated/wideresnet.py
